# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ascoor API
  version: 0.1.0
  description: API contract for the Ascoor 3D Design editor backend.
servers:
  - url: http://localhost:3100
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /design-jobs:
    post:
      operationId: createDesignJob
      tags:
        - design-jobs
      summary: Queue a design generation job from a raw user prompt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignJobRequest'
      responses:
        '202':
          description: Design job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignJobResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /designs:
    post:
      operationId: createDesign
      tags:
        - designs
      summary: Create a new project design.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignRequest'
      responses:
        '201':
          description: Design created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /projects:
    post:
      operationId: createProject
      tags:
        - projects
      summary: Create a new project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: listProjects
      tags:
        - projects
      summary: List projects for the authenticated user.
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of projects to return.
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - name: cursor
          in: query
          required: false
          description: Cursor returned by the previous paginated response.
          schema:
            type: string
        - name: q
          in: query
          required: false
          description: Case-insensitive search text for project names.
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: Projects retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    CreateDesignJobRequest:
      type: object
      required:
        - projectId
        - userPrompt
      properties:
        projectId:
          type: string
          description: Project ID that will own the generated design.
        userPrompt:
          type: string
          minLength: 1
          maxLength: 3000
          description: Raw user input prompt before AI template compilation.
    DesignJobResponse:
      type: object
      required:
        - designJobId
        - status
        - userPrompt
      properties:
        designJobId:
          type: string
          description: Design job ID.
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        userPrompt:
          type: string
          nullable: true
          description: Raw user input prompt.
        message:
          type: string
          nullable: true
          description: Assistant message for the user when completed.
        title:
          type: string
          nullable: true
          description: Suggested display name for the generated design.
        designId:
          type: string
          nullable: true
          description: ID of the created design when code is generated.
        errorMessage:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
          description: Machine-readable failure code.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 300
...
    ProjectResponse:
      allOf:
        - $ref: '#/components/schemas/ProjectResponseData'
    ProjectResponseData:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectsResponse:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponseData'
        nextCursor:
          type: string
          nullable: true
...
    CreateDesignRequest:
      type: object
      required:
        - projectId
        - displayName
        - type
      properties:
        projectId:
          type: string
        displayName:
          type: string
          minLength: 1
          maxLength: 300
        type:
          type: string
          enum: [studio_ts]
    DesignResponse:
      type: object
      required:
        - id
        - projectId
        - displayName
        - type
        - assetStatus
        - updatedAt
      properties:
        id:
          type: string
        projectId:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [studio_ts]
        assetUriTs:
          type: string
          nullable: true
        assetStatus:
          type: string
          enum: [queued, processing, succeeded, failed]
        assetError:
          type: string
          nullable: true
        editedAssetUriGlb:
          type: string
          nullable: true
        editedAssetUpdatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...