# web/shared/api/generated/schemas/createDesignRequest.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Ascoor API
 * API contract for the Ascoor 3D Design editor backend.
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDesignRequestType } from './createDesignRequestType';

export interface CreateDesignRequest {
  projectId: string;
  /**
   * @minLength 1
   * @maxLength 300
   */
  displayName: string;
  type: CreateDesignRequestType;
}